Inline SearchReflector::new in build(); drop the search wrappers
Each build() already constructed its announcement/advertisement handler
inline, but the search handler went through a reflector()/search_reflector()
wrapper. Construct SearchReflector::new directly in both build()s for
uniformity: SSDP's dial->DialRewrite/NoRewrite reply factory is now a `let
make_reply` above the register; WSD's trivial NoRewrite factory inlines.

The SSDP search glue (verdict + MX window) was all that remained in
ssdp/search.rs — too thin for its own module, so it moves into ssdp.rs and
the file is deleted.
